debug

+++ pyramid-mos/pyramid_train.py
@@ -68,9 +68,9 @@ for ep in range(1, epochs+1):
     for batch in loader:
         aud = batch['aud'].to(device)
         lens = batch['lens']
-        scores = batch['score'].to(device).unsqueeze(-1)
-        pred_scores = model(aud, lens)
-        batch_loss = criterion(pred_scores, scores).float()
+        scores = batch['score'].to(device).unsqueeze(-1).float()
+        pred_scores = model(aud, lens).float()
+        batch_loss = criterion(pred_scores, scores)
         optimizer.zero_grad()
         batch_loss.backward()
         optimizer.step()
